requirements.txt
# Core
openai>=1.12.0
python-dotenv>=1.0.0

# Orchestration
langgraph>=0.0.40
langchain>=0.1.16   # langgraph depends on langchain

# Frontend
streamlit>=1.33.0

# Vector DB
faiss-cpu>=1.7.4
numpy>=1.26.0

# API calls
requests>=2.31.0